Modularized Thumby API, lower import times (TR4)
Thumby API components are now modular, and there is now a brief overclock when loading the components, which drastically lowered the overall import time. In addition, the Sprite class now has a built-in bounding box test.
