+++ ThumbyGames/lib/thumbyLinkBase.py
@@ -1,0 +1,130 @@
+# Thumby link base
+
+from machine import Pin, UART
+
+class LinkClass:
+    def __init__(self):
+        self.initialized = False
+    
+    
+    @micropython.native
+    def init(self):
+        self.rxPin = Pin(1, Pin.IN)
+        self.uart = UART(0, baudrate=115200, rx=self.rxPin, tx=Pin(0, Pin.OUT), timeout=0, txbuf=515, rxbuf=515)
+        Pin(2, Pin.OUT).value(1)
+        
+        self.uart.read()
+        self.initialized = True
+        
+        self.sent = False
+        self.timeAtLastSend = 0
+        
+        self.timeout = 100
+    
+    
+    @micropython.native
+    def send(self, data):
+        if self.initialized != True:
+            self.init()
+        
+        uart = self.uart
+        if self.sent == False and self.rxPin.value() == 1:
+            
+            dataLength = len(data)
+            
+            if dataLength < 0 or dataLength > 512:
+                raise Exception("Link message size out of bounds: " + str(dataLength))
+            
+            packetLength = dataLength + 3
+            checksum = 0
+            
+            for b in data:
+                checksum ^= b
+            
+            anyAfter = uart.any() + packetLength
+            uart.write(bytearray([(dataLength >> 8) & 0xff, dataLength & 0xff, checksum]))
+            uart.write(data)
+            
+            # Wait to receive echo data, make sure to timeout in case of some collision
+            t0 = ticks_ms()
+            curAny = 0
+            lastAny = 0
+            while True:
+                curAny = uart.any()
+                
+                if curAny >= anyAfter:
+                    break
+                elif curAny != lastAny:
+                    # Reset timeout to accommodate longer strings of data
+                    t0 = ticks_ms()
+            
+                if ticks_diff(ticks_ms(), t0) >= self.timeout:
+                    break
+                
+                lastAny = curAny
+            
+            uart.read(anyAfter)
+            
+            self.timeAtLastSend = ticks_ms()
+            self.sent = True
+            return True
+        
+        elif self.sent == True and ticks_diff(ticks_ms(), self.timeAtLastSend) > self.timeout:
+            self.sent = False
+        
+        return False
+        
+    
+    @micropython.native
+    def receive(self):
+        if self.initialized != True:
+            self.init()
+        
+        uart = self.uart
+        
+        if uart.any() >= 3:
+            headerBytes = uart.read(3)
+            
+            receivedDataLength = (headerBytes[0] << 8) + headerBytes[1]
+            
+            # Wait for all data or timeout
+            t0 = ticks_ms()
+            curAny = 0
+            lastAny = 0
+            while True:
+                curAny = uart.any()
+                
+                if curAny >= receivedDataLength:
+                    break
+                elif curAny != lastAny:
+                    # Reset timeout to accommodate longer strings of data
+                    t0 = ticks_ms()
+            
+                if ticks_diff(ticks_ms(), t0) >= self.timeout:
+                    self.sent = False
+                    return None
+                
+                lastAny = curAny
+
+            receivedData = uart.read(receivedDataLength)
+            receivedChecksum = headerBytes[2]
+            
+            if uart.any() == 0:
+                self.sent = False
+            
+            if len(receivedData) == receivedDataLength:
+                checksum = 0
+                for b in receivedData:
+                    checksum ^= b
+    
+                if checksum == receivedChecksum:
+                    return receivedData
+                else:
+                    uart.read()
+            else:
+                uart.read()
+        elif self.sent == True and ticks_diff(ticks_ms(), self.timeAtLastSend) > self.timeout:
+            self.sent = False
+
+# Link instantiation
+link = LinkClass()
